Show widgets before acquiring a style context.
This fixes a crash with newer versions of PyGObject and/or GTK+. Note
that the helper widgets are not actually shown on screen, since they
don't have a parent window, but calling 'show' finalizes something that
allows one to get a style context.

https://bugzilla.gnome.org/show_bug.cgi?id=687513

Also, use Python attributes instead of 'get_data' and 'set_data', which
are deprecated.